organize libs into new folder
