+++ lib/asg/stringTools.lua
